update retrogolib and adapt to API changes

+++ pkg/bus/mapper.go
@@ -4,7 +4,7 @@ import "github.com/retroenv/retrogolib/arch/system/nes/cartridge"
 
 // MapperState contains the current state of the mapper.
 type MapperState struct {
-	ID   byte   `json:"id"`
+	ID   uint16 `json:"id"`
 	Name string `json:"name"`
 
 	ChrWindows []int `json:"chrWindows"`
